requirements.txt
Django~=5.0
djangorestframework~=3.15
django-filter~=24.2
django-cors-headers~=4.4
celery~=5.4
django-celery-results
boto3~=1.34
python-dotenv~=1.0
django-crontab~=0.7
psycopg2~=2.9
uvicorn[standard]~=0.30
drf-spectacular~=0.27
gunicorn


# Testing
pytest~=8.2
pytest-django~=4.8
pytest-celery~=1.0
pytest-cov~=5.0
Faker~=25.0

# Linting & formatting
black~=24.8
flake8~=7.1
isort~=5.13

# Typing
mypy~=1.13
django-stubs~=5.0
djangorestframework-stubs~=3.15
mypy-extensions

# Local dev DB convenience
psycopg2-binary~=2.9

# Celery monitoring (usually dev/tooling)
flower~=2.0

